NI-DAQmx.vi wired internally to use PFI0, Falling Edge, if Start Triggering is in use; trigger variables removed from Meas_Global.vi.
Task name is (again!) NOT held in Shift Register, or persisted in global, as this is NOT required:
The /task/ is updated by DAQmx VIs, but its /name/ is unaffected- you can see it being passed unchanged inside the DAQmx VIs. The exception would be if you did not Create a Named Task at the outset, and allowed the name to be assigned automatically, in which case you of course would need to use the name provided by the VI creating the task.
Added warning dialogues for unsupported trigger modes.
Reduced Rendezvous timeout in case of 3x failed Resets to 1000ms.


git-svn-id: https://svn.code.sf.net/p/enflo-sw/code/trunk@8122 5fe645c2-b4dc-44a5-aa5e-95b2f0996ca2